Feature/INTEG-86: Add listAccountSummaries Endpoint and Handle API errors (#2508)
* Reorganized file structure for more accurately represent the pages being displayed in the UI
- this includes test files as well

* Replaced custom css with appropriate Forma36 components
- custom css kept whenever appropriate

* Slight update to the UI/UX where the editing/installation is in the same card as the status card
- the view is toggled based on the edit and cancel button
