+++ apps/google-analytics-4/frontend/src/components/config-screen/HomeAnalytics.styles.ts
@@ -1,0 +1,27 @@
+import { css } from 'emotion';
+import tokens from '@contentful/f36-tokens';
+import { googleAnalyticsBrand } from 'components/common/GoogleAnalyticsIcon';
+
+export const styles = {
+  body: css({
+    height: 'auto',
+    minHeight: '65vh',
+    margin: '0 auto',
+    marginTop: tokens.spacingXl,
+    padding: `${tokens.spacingXl} ${tokens.spacing2Xl}`,
+    maxWidth: tokens.contentWidthText,
+    backgroundColor: tokens.colorWhite,
+    zIndex: 2,
+    boxShadow: '0px 0px 20px rgba(0, 0, 0, 0.1)',
+    borderRadius: '2px',
+  }),
+  background: css({
+    display: 'block',
+    position: 'absolute',
+    zIndex: -1,
+    top: 0,
+    width: '100%',
+    height: '300px',
+    backgroundColor: googleAnalyticsBrand.primaryColor,
+  }),
+};
