+++ apps/google-analytics-4/frontend/src/components/config-screen/configuration/ConfigurationPage.spec.tsx
@@ -1,0 +1,12 @@
+import { act, render, screen } from '@testing-library/react';
+import ConfigurationPage from './ConfigurationPage';
+
+describe('Config Screen component (not installed)', () => {
+  it('can render the about section', async () => {
+    await act(async () => {
+      render(<ConfigurationPage/>);
+    });
+
+    expect(screen.getByText('Configuration')).toBeInTheDocument();
+  });
+});
