+++ apps/google-analytics-4/frontend/src/components/config-screen/configuration/ConfigurationPage.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Stack, Box, Subheading, Paragraph } from '@contentful/f36-components'
+
+export default function ConfigurationPage() {
+  return (
+    <Stack spacing='spacingL' >
+      <Box>
+        <Subheading>Configuration</Subheading>
+        <Paragraph>Configure your Google Analytics app installation.</Paragraph>
+      </Box>
+    </Stack>
+  )
+}
