+++ apps/google-analytics-4/frontend/src/components/config-screen/header/AboutSection.spec.tsx
@@ -1,0 +1,12 @@
+import AboutSection from "./AboutSection";
+import { act, render, screen } from '@testing-library/react';
+
+describe('About header for the Analytics Config Page', () => {
+  it('can render the about section', async () => {
+    await act(async () => {
+      render(<AboutSection />);
+    });
+
+    expect(screen.getByText('About Google Analytics for Contentful')).toBeInTheDocument();
+  });
+});
