+++ apps/google-analytics-4/frontend/src/components/config-screen/header/AboutSection.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Box, Heading, Paragraph } from '@contentful/f36-components'
+
+export default function AboutSection() {
+  return (
+    <Box>
+      <Heading>About Google Analytics for Contentful</Heading>
+      <Paragraph>
+        The Google Analytics app displays real-time page-based analytics data from your
+        organization’s Google Analytics properties alongside relevant content entries.
+      </Paragraph>
+    </Box>
+  )
+}
