+++ apps/google-analytics-4/frontend/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "@tsconfig/create-react-app/tsconfig.json",
   "compilerOptions": {
+    "baseUrl": "src"
   },
   "include": [
     "src"
